cj/bin/every10min_fx.bash

+++ bin/every10min_fx.bash
@@ -5,31 +5,43 @@
 # I use this script to copy data out of the DB into some partials.
 # Then, it uses git to copy those partials to the Rails site.
 
-cd /pt/s/rl/cj4svm/
-. ./.cj
+# If the /tmp/script_e10_fx_busy.txt exists, I should exit
 
-# Now for Forex,
-# use expdp to copy data from active-fx-db into local-db:
-ssh z /pt/s/rl/cj4svm/bin/expdp_fx.bash
+if [ -e "/tmp/script_e10_fx_busy.txt" ]; then
+   echo The /tmp/script_e10_fx_busy.txt exists.
+   echo If the /tmp/script_e10_fx_busy.txt exists, I should exit.
+   exit
+else
+  date > /tmp/script_e10_fx_busy.txt
+  cd /pt/s/rl/cj4svm/
+  . ./.cj
 
-rsync z:dpdump/fx.dpdmp ~/dpdump/
-impdp trade/t table_exists_action=replace dumpfile=fx.dpdmp
+  # Now for Forex,
+  # use expdp to copy data from active-fx-db into local-db:
+  ssh z /pt/s/rl/cj4svm/bin/expdp_fx.bash
 
-# Copy data out of the DB into some partials:
-cd /pt/s/rl/cj4svm/predictions/fx_new/
-./index_spec.bash
+  rsync z:dpdump/fx.dpdmp ~/dpdump/
+  impdp trade/t table_exists_action=replace dumpfile=fx.dpdmp
 
-# Now copy the new data to the Rails site:
+  # Copy data out of the DB into some partials:
+  cd /pt/s/rl/cj4svm/predictions/fx_new/
+  ./index_spec.bash
 
-set -x
+  # Now copy the new data to the Rails site:
 
-cd /pt/s/rl/svm/
-git add .
-git commit -a -v -m every10min.bash-commit
-git push origin master
-git push heroku master
+  set -x
 
-# Now, pull the new data into the Varnish-cache at the server:
-/pt/s/rl/cj4svm/bin/wgetit.bash
+  cd /pt/s/rl/svm/
+  git add .
+  git commit -a -v -m every10min.bash-commit
+  git push origin master
+  git push heroku master
+
+  # Now, pull the new data into the Varnish-cache at the server:
+  /pt/s/rl/cj4svm/bin/wgetit.bash
+
+  rm -f  /tmp/script_e10_fx_busy.txt
+fi
 
 exit 0
+
